+++ app/source/runtime/infer/private/InferOCR.h
@@ -1,0 +1,1 @@
+#pragma once
