+++ scripts/build-release.bat
@@ -1,2 +1,2 @@
-xmake f -c -p windows -a x64 -k binary -m release --runtimes="MT" -vDy &&^
+xmake f -c -p windows -a x64 -k binary -m release --runtimes="MD" -vDy &&^
 xmake build
