+++ scripts/setupdev-vs.bat
@@ -1,2 +1,2 @@
-xmake f -c -p windows -a x64 -m releasedbg --runtimes="MDd" -vDy &&^
+xmake f -c -p windows -a x64 -m "releasedbg" --runtimes="MDd" -vDy &&^
 xmake project -k vsxmake -m "releasedbg" -a "x64" -P . -y -vD
