+++ scripts/setuprel-vs.bat
@@ -1,0 +1,2 @@
+xmake f -c -p windows -a x64 -m "release" --runtimes="MD" -vDy &&^
+xmake project -k vsxmake -m "release" -a "x64" -P . -y -vD
